refactor(test): split trow snap fixture
